requirements.txt
Django==3.2
djangorestframework==3.12.2
psycopg2-binary==2.9.3 
paho-mqtt==1.5.1
gunicorn
flake8==3.9.0

